re配置完全版なおゲームの進行は厳しい模様

+++ Assets/FudaScript/MoveFuda.cs
@@ -9,7 +9,7 @@ public class MoveFuda : MonoBehaviour {
       if(GameObject.FindWithTag("check") == null ){
         if(this.tag == "playerfuda"){
           this.screenPoint = Camera.main.WorldToScreenPoint(transform.position);
-          this.offset = transform.position - Camera.main.ScreenToWorldPoint(new Vector3(Input.mousePosition.x, Input.mousePosition.y, screenPoint.z));
+          this.offset = transform.position - Camera.main.ScreenToWorldPoint(new Vector3((int)Input.mousePosition.x, (int)Input.mousePosition.y, (int)screenPoint.z));
         }
       }
     }
@@ -17,7 +17,7 @@ public class MoveFuda : MonoBehaviour {
     void OnMouseDrag() {
       if(GameObject.FindWithTag("check") == null ){
         if(this.tag == "playerfuda"){
-          Vector3 currentScreenPoint = new Vector3(Input.mousePosition.x, Input.mousePosition.y, screenPoint.z);
+          Vector3 currentScreenPoint = new Vector3((int)Input.mousePosition.x, (int)Input.mousePosition.y, (int)screenPoint.z);
           Vector3 currentPosition = Camera.main.ScreenToWorldPoint(currentScreenPoint) + this.offset;
           transform.position=(new Vector3(Mathf.Clamp(currentPosition.x,0,112),Mathf.Clamp(currentPosition.y,0,0),Mathf.Clamp(currentPosition.z,-40,-20)));
           //transform.position = currentPosition;
